+++ Serilog.Sinks.Console.Themes.UnitTests/TrueColorTests.cs
@@ -29,6 +29,23 @@ internal sealed class TrueColorTests
     }
 
     [Test]
+    public void Background_ConsoleColor_matches_Color_FromName()
+    {
+        var cc = ConsoleColor.DarkGreen;
+        var c = Color.FromName(cc.ToString());
+        var expected = $"\x1b[48;2;{c.R};{c.G};{c.B}m";
+        TrueColor.Background(cc).Should().Be(expected);
+    }
+
+    [Test]
+    public void Foreground_Color_and_Background_Color_emit_true_color_sgr()
+    {
+        var c = Color.FromArgb(11, 22, 33);
+        TrueColor.Foreground(c).Should().Be($"\x1b[38;2;{c.R};{c.G};{c.B}m");
+        TrueColor.Background(c).Should().Be($"\x1b[48;2;{c.R};{c.G};{c.B}m");
+    }
+
+    [Test]
     public void BoldForegroundBackground_KnownColor_prefixes_bold_and_combines_fg_bg()
     {
         var fg = Color.FromKnownColor(KnownColor.White);