+++ Serilog.Sinks.Console.Themes/AssemblyInfo.cs
@@ -1,1 +1,1 @@
-[assembly: InternalsVisibleTo("Serilog.Sinks.Console.Themes.UnitTests")]
+﻿[assembly: InternalsVisibleTo("Serilog.Sinks.Console.Themes.UnitTests")]